Optimize and Enhance
### 1. Improved Type Definitions
The code now features more explicit and detailed TypeScript type definitions for the `PaginationProps`. This not only enhances code readability but also provides stronger type checking, helping to catch potential bugs or type errors during development.

### 2. Optimized Code for Generating Button Numbers Array
The generation of the `buttonNumbers` array has been optimized. Instead of creating an empty array and pushing numbers into it in a loop, the new code uses `Array.from` with a length derived from the start and end page numbers. This approach is cleaner, more idiomatic, and potentially more efficient.

### 3. Better Use of `useCallback` Hook
The `handlePageChange` function is now wrapped in the `useCallback` hook. This hook returns a memoized version of the function that only changes if one of the dependencies has changed. This can potentially improve performance by avoiding unnecessary re-renders or function recreations.

### 4. Enhanced Customization Options for the Button Component
The updated code now allows for additional customization of the `Button` component. The `className` prop was added to the `Button` component, providing the ability to add a custom CSS class to the pagination buttons. 

### 5. Defined Constants for Default Values
Constants for default values have been defined to improve consistency and readability.

### 6. Added `React.memo` to the Button Component
This addition can lead to potential performance optimizations by preventing unnecessary re-renders of the `Button` component if its props haven't changed.

**Note:** The `className` prop was originally added to the Pagination component as part of a separate pull request ([#1](https://github.com/BankkRoll/nextjs-pagination/pull/1)) by @nkp1111. Credit to them for this contribution.